requirement.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.0
Django==5.1.3
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.6
pycparser==2.22
PyJWT==2.10.0
python-dotenv==1.0.1
sqlparse==0.5.2
